backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
openpyxl==3.1.5
pydantic==2.10.4
pydantic-settings==2.7.0
